Move handlers to a different file
